requirements.txt
pytest==9.0.1
python-dotenv==1.2.1
selenium==4.38.0
